PGA Moneyball
Colombia University Professor Mark Broadie is one of many who have studied the statistics of the game and have come to the conclusion that driving distance is a better predictor of winnings on Tour than putting ability. Are there any single measures which better predict winnings on tour than driving distance? Can we build a reliable model which accurately predicts winnings on Tour?